Remove save on main "Choose" screen.
This commit fixes the app so that it no longer saves items on the main
"Choose" screen, which was an error in requirements analysis. Items are
still saved on the "Recent" selection screen. This fixes GitHub issue #7.

This next bit may change as we learn more about the original app. Now,
choices are saved to working memory, but not to device storage on the
"Choose" screen. They remain as new choices are added, and after a choice
is made, but will disappear when the app is terminated or
when navigating all the way back to the main menu.