Prevent simultaneous press of multiple buttons
This commit fixes strange navigation issues and problems with
the navigation bar titles that occurred when multiple buttons were
pressed simultaneously.

Now, only one button at a time on the main menu can receive a press event.
Similar changes were made to buttons in the navigation bars.

Prior to these fixes, pressing multiple buttons led to
two navigation actions being executed simultaneously, confusing navigation.
When the buttons were in the navigation bar, the titles of each screen would also be
overlaid on each other and error log messages were triggered:

2016-05-23 06:37:18.348 Re-Resolver[3854:1558530] nested push animation can result in corrupted navigation bar
2016-05-23 06:37:18.725 Re-Resolver[3854:1558530] Finishing up a navigation transition in an unexpected state. Navigation Bar subview tree might get corrupted.

Testing of prior versions of this app also occasionally triggered a condition where
pressing a back button on the navigation bar did not move back. It's not known - but hoped -
that the issue is also corrected with these fixes.