Fix issue with repeated quick press on "+" button.
This fixes another issue that could have potentially corrupted
navigation. In the "Choose" function, if you quickly and repeatedly
pressed the touch screen in the "+" and "Recent" buttons in the toolbar
area immediately after picking "Choose" from the main menu, this led
to multiple "Add Choice" screens being stacked on top of each other.
This corrects the issue by disabling the "+" button soon after it is touched.