Update "Ask" screen with genuine Resolver answers
These answers for the Ask screen were found by pressing the button
on the Ask screen in the original app over and over again.

They replace the Magic 8-ball answers. There are 20 and I believe that
to be a complete set, but there is a possibility that some are missing.