Change "Recent" screen to return directly to "Choose" list
The "Recent" screen previously returned to the "New Choice"
screen so that the selected recent entry could be edited before
it was added to the list of choices.

Now, when a Recent item is selected, the "New Choice" screen
is bypassed and the item is added directly to the "Choose" screen.
This is the behavior of the classic app.

There may be an opportunity for refactoring after this change. The
"ChooseViewController" is now serving as the NewChoiceDelegate and
the RecentItemDelegate, to handle selections from the "New Choice" screen
and the "Recent Item" screen. The "New Choice" screen handles setting
the previous screen, the ChooseViewController, as the RecentItemDelegate.
This seems strange and perhaps both delegates could be replaced with an
unwind segue back to ChooseViewController.