Explicitly ignore return value from popViewController
With the Swift 3.0 update, the compiler is warning about unused
return values from popViewController. The popViewController
method returns the controller that was just popped
from the stack.

I don't have a use for the return value at this time, so I'm
ignoring it by assigning it to the special _ variable.