Use original image assets if the phone is using English.
The original button images have "Decide", "Choose", "Ask"
as part of the image. If the language is English, we use the
original image assets for authenticity. If the app is using a language
other than English, use blank image assets (without words) and
place text strings from our localization file into labels on the button.