Increment version to 1.6 and build to 6
The previous archive marked as version 1.1 and build 5
was automatically rejected by the App Store. The previous
version of the software was 1.05, and some component in
the process interpreted version number 1.1, the original number
for the new version, to be lower than 1.05, thinking that 1.05 was
equivalent to 1.5. I may have broken a convention in software
versioning by placing more than 1 digit after a period.

So the version has been renamed from 1.1 to 1.6, and the build
number has also been incremented.