Update navigation to modern appearance
This removes the custom back button look for the navigation bars,
which was in place so that the app looked like the original Resolver.
This also adds large, clear titles for iOS 11 users.

These changes were spurred on by problems with the navigation
bar, in particular the back buttons, under iOS 11.2. The behavior of
this part of the IOS SDK seems to change frequently, causing a lot
of maintenance for this feature. It's undecided whether the customization
will be added back to a future version of Re-Resolver.

Fixes #35 in GitHub.