Add privacy policy
As of October 3, 2018, updates to apps in the App Store require a privacy
policy. This policy covers data collection, data sharing, and data
retention, as required by App Store guidelines. ReResolver doesn't have
access to or make use of any data that isn't needed to perform core
functionality, and now that's made explicit in the policy.

I'm unsure, without having researched other examples, if this draft is too
detailed or not detailed enough, so this may see some updates before the
next version of the app is published.