Automatically size color cells to support larger screen sizes
This fixes a usability issue with the app on the 6.1 inch XR and 6.5 inch XS Max
released in 2018. Color cells were appearing in more than one
row for these large screens, which may have been confusing for a collection
which allows only horizontal swiping.

This fix was tested only in the simulator, since at present there is no budget
for new devices.