Adjust top of instructions screen to fix large title problem
This changes the instructions text margins again so that
the top of the text will NOT scroll under the navigation bar,
but will scroll right up to the bottom edge of the navigation bar instead.

The previously added effect was causing a jumping effect in the
navigation bar from large to small title before the user touched
the device.