Add initial iPad support
This does most of the work of allowing the app to run
on the iPad without using iPhone simulator mode. The
app can run in full screen, or in Slide Over or Split
View multitasking modes.

- Support all portrait and landscape orientations
- Modify color selector to adjust size of cells based on
  window size, instead of the size of the entire screen
- Adjust build settings for main app and the test targets
  in order to support the iPad device family.

The deployment target iOS version settings for the testing
targets and the main app were also inconsistent with each
other, and have been adjusted to 10.0.

This work would also be necessary to release
a Catalyst version of ReResolver that would run directly
on MacOS, however, I have no plans to release
a Mac version of ReResolver because nobody
has asked for that.