Opt out of new modal presentation style
This changes the color selection screen to use the
old fullscreen presentation style instead of the new
iOS 13 presentation style that can be dismissed with
a downward swipe.

This serves as a quick fix for issues #46, #47,
and #49 in the GitHub project, all related to the
color selection screen in Xcode 11/iOS 13.

Releasing an app with this quick fix, built with
Xcode 11, with also fix issue #45, the dark
navigation bar titles in iOS 13.

I prefer the new presentation style, but the app
requires some updates to work properly with the
new style and I may not be able to make those
updates in time for the iOS 13 release date due
to likely overtime at my job this week and next.