Fix issue with missing navigation bar after using color selector
With the new card based modal presentation style in iOS13,
the viewWillDisappear event was not being triggered for the
ResolverMenuViewController when the color selector
appeared, as the main menu now remains partially on
screen under the color selector. This caused the
navigation bar to always be hidden after the user
selected a color.

This change removes code to work around a no longer
existent issue with the Resolver logo dropping down,
thus avoiding the problem with the disappearing
navigation bar.