Fix select color crash in Mac build running on Big Sur beta.
I was experimenting with Big Sur and tried a Mac Catalyst
build of the ReResolver app. The app crashed when bringing up
the "Select Color" window.

Even though the Mac version of ReResolver isn't released or
supported, it seemed like a good idea to fix this bug, because
it was caused by bad code which assumed that an array
was not empty.

This is the quick fix - tentatively access the array through the "first"
member instead of index 0. I may look into this more later, though,
as it's not immediately clear to me why the array of selected cells
was empty in this case, and why it works on other builds and platforms.